% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  connectionDetails,
  usePackageCohorts = T,
  newTargetCohortId = NULL,
  newOutcomeCohortId = NULL,
  newCohortDatabaseSchema = NULL,
  newCohortTable = NULL,
  cdmDatabaseSchema,
  cdmDatabaseName = "friendly database name",
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = cohortDatabaseSchema,
  sampleSize = NULL,
  riskWindowStart = 1,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  firstExposureOnly = F,
  removeSubjectsWithPriorOutcome = F,
  priorOutcomeLookback = 99999,
  requireTimeAtRisk = F,
  minTimeAtRisk = 1,
  includeAllOutcomes = T,
  endDay = -1,
  outputFolder,
  createCohorts = F,
  runAnalyses = F,
  viewShiny = F,
  shinyAnalysisId = 1,
  packageResults = F,
  minCellCount = 10,
  verbosity = "INFO",
  cdmVersion = 5
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{usePackageCohorts}{Whether to use the T and O cohorts in the package}

\item{newTargetCohortId}{The cohort id of a new target cohort (must be generated already)}

\item{newOutcomeCohortId}{The cohort id of a new outcome (must be generated already)}

\item{newCohortDatabaseSchema}{Schema name where the new cohorts are stored. Note that for SQL Server, 
this should include both the database and schema name, for example 'cdm_data.dbo'.}

\item{newCohortTable}{The name of the table that contains the target population and outcome
cohorts.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cdmDatabaseName}{Shareable name of the database}

\item{cohortDatabaseSchema}{Schema name where the cohorts and covariates are created. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the target, outcome and variable cohorts used in this study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{sampleSize}{How many patients to sample from the target population}

\item{riskWindowStart}{The start of the risk window (in days) relative to the startAnchor.}

\item{startAnchor}{The anchor point for the start of the risk window. Can be "cohort start" or "cohort end".}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the endAnchor parameter}

\item{endAnchor}{The anchor point for the end of the risk window. Can be "cohort start" or "cohort end".}

\item{firstExposureOnly}{Should only the first exposure per subject be included? Note that this is typically done in the createStudyPopulation function,}

\item{removeSubjectsWithPriorOutcome}{Remove subjects that have the outcome prior to the risk window start?}

\item{priorOutcomeLookback}{How many days should we look back when identifying prior outcomes?}

\item{requireTimeAtRisk}{Should subject without time at risk be removed?}

\item{minTimeAtRisk}{The minimum number of days at risk required to be included}

\item{includeAllOutcomes}{(binary) indicating whether to include people with outcomes who are not observed for the whole at risk period}

\item{endDay}{The end day relative to index for the custom covariates (default is -1)}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{createCohorts}{Create the cohortTable table with the cohorts}

\item{runAnalyses}{Run the model development}

\item{viewShiny}{View the results as a shiny app}

\item{shinyAnalysisId}{The analysisId corresponding to the result you want to view}

\item{packageResults}{Should results be packaged for later sharing?}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included 
in packaged results.}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{cdmVersion}{The version of the common data model}
}
\description{
Execute the Study
}
\details{
This function executes the CovidVulnerabilityIndex Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        cdmDatabaseName = 'shareable name of the database'
        cohortDatabaseSchema = "study_results",
        cohortTable = "cohort",
        outcomeId = 1,
        oracleTempSchema = NULL,
        riskWindowStart = 1,
        startAnchor = 'cohort start',
        riskWindowEnd = 365,
        endAnchor = 'cohort start',
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        runAnalyses = T,
        viewShiny = F,
        shinyAnalysisId = 1,
        packageResults = F,
        minCellCount = 10,
        verbosity = "INFO",
        cdmVersion = 5)
}

}
